Refactor custom preset logic & add support for aacaudio codec
