+++ azure-cli2017.pyproj
@@ -140,6 +140,9 @@
     <Compile Include="command_modules\azure-cli-advisor\azure_bdist_wheel.py" />
     <Compile Include="command_modules\azure-cli-advisor\setup.py" />
     <Compile Include="command_modules\azure-cli-ams\azure\cli\command_modules\ams\commands.py" />
+    <Compile Include="command_modules\azure-cli-ams\azure\cli\command_modules\ams\_sdk_utils.py">
+      <SubType>Code</SubType>
+    </Compile>
     <Compile Include="command_modules\azure-cli-ams\azure\cli\command_modules\ams\operations\account.py" />
     <Compile Include="command_modules\azure-cli-ams\azure\cli\command_modules\ams\operations\asset.py" />
     <Compile Include="command_modules\azure-cli-ams\azure\cli\command_modules\ams\operations\job.py" />
@@ -765,6 +768,7 @@
     <Folder Include="command_modules\azure-cli-ams\azure\cli\command_modules\ams\operations\" />
     <Folder Include="command_modules\azure-cli-ams\azure\cli\command_modules\ams\tests\" />
     <Folder Include="command_modules\azure-cli-ams\azure\cli\command_modules\ams\tests\latest\" />
+    <Folder Include="command_modules\azure-cli-ams\azure\cli\command_modules\ams\tests\latest\data\" />
     <Folder Include="command_modules\azure-cli-ams\azure\cli\command_modules\ams\tests\latest\recordings\" />
     <Folder Include="command_modules\azure-cli-appservice\" />
     <Folder Include="command_modules\azure-cli-appservice\azure\" />
@@ -1068,6 +1072,8 @@
     <Content Include="command_modules\azure-cli-advisor\MANIFEST.in" />
     <Content Include="command_modules\azure-cli-advisor\README.rst" />
     <Content Include="command_modules\azure-cli-advisor\setup.cfg" />
+    <Content Include="command_modules\azure-cli-ams\azure\cli\command_modules\ams\tests\latest\data\customPreset.json" />
+    <Content Include="command_modules\azure-cli-ams\azure\cli\command_modules\ams\tests\latest\data\invalidCustomPreset.json" />
     <Content Include="command_modules\azure-cli-ams\azure\cli\command_modules\ams\tests\latest\recordings\test_ams_asset.yaml" />
     <Content Include="command_modules\azure-cli-ams\azure\cli\command_modules\ams\tests\latest\recordings\test_ams_asset_get_sas_urls.yaml" />
     <Content Include="command_modules\azure-cli-ams\azure\cli\command_modules\ams\tests\latest\recordings\test_ams_create_show.yaml" />
