+++ src/command_modules/azure-cli-ams/azure/cli/command_modules/ams/tests/latest/data/customPreset.json
@@ -47,7 +47,7 @@
             "Level": "auto",
             "Bitrate": 1045,
             "MaxBitrate": 1045,
-            "BufferWindow": "00:00:05",
+            "BufferWindow": "00:00:04",
             "ReferenceFrames": 3,
             "EntropyMode": "Cavlc",
             "AdaptiveBFrame": true,
